requirements.txt
gTTS==2.0.1
numpy==1.15.4
python-vlc==3.0.4106
